\name{arpautils-package}
\alias{arpautils-package}
\alias{arpautils}
\docType{package}
\title{
The Arpa ER Utils Package
}
\description{
R utility functions for Arpa ER
}
\details{
\tabular{ll}{
Package: \tab arpautils\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-31\cr
License: \tab What license is it under? TO BE DEFINED\cr
}
Funzioni utili per dati QA
}
\author{
G.Bonafe' <gbonafe@arpa.emr.it>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
